Move DummyLogo to components
