+++ packages/components/src/components/DummyLogo/DummyLogo.js
