+++ packages/components/src/components/DummyLogo/index.js
