+++ packages/components/src/components/Footer/Footer.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select } from '@storybook/addon-knobs';
-import { DummyLogo } from '../../helpers';
+import DummyLogo from '../DummyLogo';
 import FlexList from '../FlexList';
 import { GitHubIcon, TwitterIcon } from '../../icons';
 import SmartLink from '../SmartLink';
