+++ packages/components/src/components/Header/Header.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select } from '@storybook/addon-knobs';
 import { Box } from '@theme-ui/components';
-import { DummyLogo } from '../../helpers';
+import DummyLogo from '../DummyLogo';
 import theme from '../../theme';
 import Header from './Header';
