+++ packages/components/src/components/index.js
@@ -1,5 +1,6 @@
 export { default as Badge } from './Badge';
 export { default as Container } from './Container';
+export { default as DummyLogo } from './DummyLogo';
 export { default as EventPreview, eventType } from './EventPreview';
 export { default as FlexList } from './FlexList';
 export { default as Footer } from './Footer';
