--- packages/components/src/helpers/index.js
@@ -1,1 +1,0 @@
-export { default as DummyLogo } from './DummyLogo';
