+++ packages/components/src/index.js
@@ -2,6 +2,7 @@ export { Box, Button, Flex, Heading, Text } from '@theme-ui/components';
 export {
   Badge,
   Container,
+  DummyLogo,
   EventPreview,
   eventType,
   FlexList,
@@ -17,4 +18,3 @@ export {
 } from './components';
 export { GitHubIcon, TwitterIcon } from './icons';
 export { default as theme } from './theme';
-export { DummyLogo } from './helpers';
